Adds tox checks
