+++ shared_test_utils/tox.ini
@@ -1,0 +1,31 @@
+# Copyright 2023 Canonical Ltd.
+# See LICENSE file for licensing details.
+[tox]
+skipsdist=True
+skip_missing_interpreters = True
+envlist = lint, static
+
+[vars]
+constants_path = {toxinidir}/integration_constants.py
+src_path = {toxinidir}/src
+all_path = {[vars]constants_path} {[vars]src_path}
+
+[testenv]
+deps = 
+    -r{toxinidir}/../test-requirements.txt
+setenv =
+    PYTHONPATH = {toxinidir}:{toxinidir}/lib:{[vars]src_path}
+
+[testenv:lint]
+description = Check code against coding style standards
+commands =
+    pflake8 {[vars]all_path}
+    isort --check-only --diff {[vars]all_path}
+    black --check --diff {[vars]all_path}
+
+[testenv:static]
+description = Run static analysis checks
+commands =
+    mypy {[vars]all_path} {posargs}
+setenv =
+    PYTHONPATH = ""
